Migrated to Network Monitor from core-foundation
