+++ core/foundation/src/main/kotlin/dev/teogor/ceres/core/foundation/FoundationGlobal.kt
@@ -17,7 +17,5 @@
 package dev.teogor.ceres.core.foundation
 
 object FoundationGlobal {
-  val networkMonitor = NetworkMonitor(
-    isOffline = true,
-  )
+  val networkMonitor = NetworkMonitor()
 }
