--- core/network/api/network.api
@@ -1,25 +1,0 @@
-public final class dev/teogor/ceres/core/network/ConnectivityManagerNetworkMonitor : dev/teogor/ceres/core/network/NetworkMonitor {
-	public fun <init> (Landroid/content/Context;)V
-	public fun isOnline ()Lkotlinx/coroutines/flow/Flow;
-}
-
-public final class dev/teogor/ceres/core/network/ConnectivityManagerNetworkMonitor_Factory : dagger/internal/Factory {
-	public fun <init> (Ljavax/inject/Provider;)V
-	public static fun create (Ljavax/inject/Provider;)Ldev/teogor/ceres/core/network/ConnectivityManagerNetworkMonitor_Factory;
-	public fun get ()Ldev/teogor/ceres/core/network/ConnectivityManagerNetworkMonitor;
-	public synthetic fun get ()Ljava/lang/Object;
-	public static fun newInstance (Landroid/content/Context;)Ldev/teogor/ceres/core/network/ConnectivityManagerNetworkMonitor;
-}
-
-public abstract interface class dev/teogor/ceres/core/network/NetworkMonitor {
-	public abstract fun isOnline ()Lkotlinx/coroutines/flow/Flow;
-}
-
-public abstract interface class dev/teogor/ceres/core/network/di/NetworkModule {
-	public abstract fun bindsNetworkMonitor (Ldev/teogor/ceres/core/network/ConnectivityManagerNetworkMonitor;)Ldev/teogor/ceres/core/network/NetworkMonitor;
-}
-
-public class hilt_aggregated_deps/_dev_teogor_ceres_core_network_di_NetworkModule {
-	public fun <init> ()V
-}
-
