--- core/network/src/main/kotlin/dev/teogor/ceres/core/network/ConnectivityManagerNetworkMonitor.kt
@@ -1,28 +1,0 @@
-/*
- * Copyright 2023 teogor (Teodor Grigor)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package dev.teogor.ceres.core.network
-
-import android.content.Context
-import dagger.hilt.android.qualifiers.ApplicationContext
-import dev.teogor.ceres.core.foundation.networkConnectivityUtils
-import javax.inject.Inject
-
-class ConnectivityManagerNetworkMonitor @Inject constructor(
-  @ApplicationContext private val context: Context,
-) : NetworkMonitor {
-  override val isOnline = context.networkConnectivityUtils().isOnline
-}
