+++ framework/core/api/core.api
@@ -3,7 +3,7 @@ public class dev/teogor/ceres/framework/core/Activity : androidx/activity/Compon
 	public field analyticsHelper Ldev/teogor/ceres/firebase/analytics/AnalyticsHelper;
 	public field crashlyticsHelper Ldev/teogor/ceres/firebase/crashlytics/CrashlyticsHelper;
 	public field lazyStats Ldagger/Lazy;
-	public field networkMonitor Ldev/teogor/ceres/core/network/NetworkMonitor;
+	public field networkMonitor Ldev/teogor/ceres/core/foundation/NetworkMonitorUtility;
 	public fun <init> ()V
 	public fun BuildNavGraph (Ldev/teogor/ceres/framework/core/model/NavGraphOptions;Landroidx/compose/runtime/Composer;I)V
 	public fun buildMenu (Ldev/teogor/ceres/framework/core/model/MenuConfig;Landroidx/compose/runtime/Composer;I)Ldev/teogor/ceres/framework/core/model/MenuConfig;
@@ -12,7 +12,7 @@ public class dev/teogor/ceres/framework/core/Activity : androidx/activity/Compon
 	public final fun getCrashlyticsHelper ()Ldev/teogor/ceres/firebase/crashlytics/CrashlyticsHelper;
 	public final fun getLazyStats ()Ldagger/Lazy;
 	public fun getNavigationItems ()Ljava/util/List;
-	public final fun getNetworkMonitor ()Ldev/teogor/ceres/core/network/NetworkMonitor;
+	public final fun getNetworkMonitor ()Ldev/teogor/ceres/core/foundation/NetworkMonitorUtility;
 	public fun handleUriVariants (Landroid/net/Uri;)Ldev/teogor/ceres/navigation/core/ScreenRoute;
 	public final fun navigateTo (Ldev/teogor/ceres/navigation/core/ScreenRoute;)V
 	protected fun onCreate (Landroid/os/Bundle;)V
@@ -23,7 +23,7 @@ public class dev/teogor/ceres/framework/core/Activity : androidx/activity/Compon
 	public final fun setCrashlyticsHelper (Ldev/teogor/ceres/firebase/crashlytics/CrashlyticsHelper;)V
 	public fun setKeepOnScreenCondition ()Z
 	public final fun setLazyStats (Ldagger/Lazy;)V
-	public final fun setNetworkMonitor (Ldev/teogor/ceres/core/network/NetworkMonitor;)V
+	public final fun setNetworkMonitor (Ldev/teogor/ceres/core/foundation/NetworkMonitorUtility;)V
 	public final fun showMenu (Z)V
 }
 
@@ -35,7 +35,7 @@ public final class dev/teogor/ceres/framework/core/Activity_MembersInjector : da
 	public static fun injectLazyStats (Ldev/teogor/ceres/framework/core/Activity;Ldagger/Lazy;)V
 	public fun injectMembers (Ldev/teogor/ceres/framework/core/Activity;)V
 	public synthetic fun injectMembers (Ljava/lang/Object;)V
-	public static fun injectNetworkMonitor (Ldev/teogor/ceres/framework/core/Activity;Ldev/teogor/ceres/core/network/NetworkMonitor;)V
+	public static fun injectNetworkMonitor (Ldev/teogor/ceres/framework/core/Activity;Ldev/teogor/ceres/core/foundation/NetworkMonitorUtility;)V
 }
 
 public class dev/teogor/ceres/framework/core/Application : android/app/Application {
@@ -74,13 +74,13 @@ public final class dev/teogor/ceres/framework/core/app/BaseActionsKt {
 }
 
 public final class dev/teogor/ceres/framework/core/app/CeresAppKt {
-	public static final fun CeresApp (Landroidx/compose/material3/windowsizeclass/WindowSizeClass;Ldev/teogor/ceres/core/network/NetworkMonitor;Ljava/util/List;Ldev/teogor/ceres/framework/core/app/CeresAppState;Ldev/teogor/ceres/framework/core/app/BaseActions;Lkotlin/jvm/functions/Function1;Lkotlin/jvm/functions/Function3;Lkotlin/jvm/functions/Function6;Landroidx/compose/runtime/Composer;II)V
+	public static final fun CeresApp (Landroidx/compose/material3/windowsizeclass/WindowSizeClass;Ldev/teogor/ceres/core/foundation/NetworkMonitorUtility;Ljava/util/List;Ldev/teogor/ceres/framework/core/app/CeresAppState;Ldev/teogor/ceres/framework/core/app/BaseActions;Lkotlin/jvm/functions/Function1;Lkotlin/jvm/functions/Function3;Lkotlin/jvm/functions/Function6;Landroidx/compose/runtime/Composer;II)V
 	public static final fun calculateInnerPadding-DH4mp8Y (Landroidx/compose/foundation/layout/PaddingValues;Landroidx/compose/ui/unit/LayoutDirection;FFFFLandroidx/compose/runtime/Composer;II)Landroidx/compose/foundation/layout/PaddingValues;
 }
 
 public final class dev/teogor/ceres/framework/core/app/CeresAppState {
 	public static final field $stable I
-	public fun <init> (Landroidx/navigation/NavHostController;Lkotlinx/coroutines/CoroutineScope;Landroidx/compose/material3/windowsizeclass/WindowSizeClass;Ldev/teogor/ceres/framework/core/screen/ScreenInfo;Ldev/teogor/ceres/core/network/NetworkMonitor;Ljava/util/List;)V
+	public fun <init> (Landroidx/navigation/NavHostController;Lkotlinx/coroutines/CoroutineScope;Landroidx/compose/material3/windowsizeclass/WindowSizeClass;Ldev/teogor/ceres/framework/core/screen/ScreenInfo;Ldev/teogor/ceres/core/foundation/NetworkMonitorUtility;Ljava/util/List;)V
 	public final fun getCoroutineScope ()Lkotlinx/coroutines/CoroutineScope;
 	public final fun getCurrentDestination (Landroidx/compose/runtime/Composer;I)Landroidx/navigation/NavDestination;
 	public final fun getFloatingButtonView ()Ldev/teogor/ceres/framework/core/screen/FloatingButtonView;
@@ -106,7 +106,7 @@ public final class dev/teogor/ceres/framework/core/app/CeresAppState {
 }
 
 public final class dev/teogor/ceres/framework/core/app/CeresAppStateKt {
-	public static final fun rememberCeresAppState (Landroidx/compose/material3/windowsizeclass/WindowSizeClass;Ldev/teogor/ceres/core/network/NetworkMonitor;Lkotlinx/coroutines/CoroutineScope;Landroidx/navigation/NavHostController;Ldev/teogor/ceres/framework/core/screen/ScreenInfo;Ljava/util/List;Landroidx/compose/runtime/Composer;II)Ldev/teogor/ceres/framework/core/app/CeresAppState;
+	public static final fun rememberCeresAppState (Landroidx/compose/material3/windowsizeclass/WindowSizeClass;Ldev/teogor/ceres/core/foundation/NetworkMonitorUtility;Lkotlinx/coroutines/CoroutineScope;Landroidx/navigation/NavHostController;Ldev/teogor/ceres/framework/core/screen/ScreenInfo;Ljava/util/List;Landroidx/compose/runtime/Composer;II)Ldev/teogor/ceres/framework/core/app/CeresAppState;
 }
 
 public final class dev/teogor/ceres/framework/core/app/ComposableSingletons$CeresAppKt {
@@ -114,17 +114,6 @@ public final class dev/teogor/ceres/framework/core/app/ComposableSingletons$Cere
 	public static field lambda-1 Lkotlin/jvm/functions/Function3;
 	public fun <init> ()V
 	public final fun getLambda-1$core_release ()Lkotlin/jvm/functions/Function3;
-}
-
-public final class dev/teogor/ceres/framework/core/compositions/NetworkConnectivity {
-	public static final field $stable I
-	public fun <init> ()V
-	public final fun isOffline ()Z
-	public final fun setOffline (Z)V
-}
-
-public final class dev/teogor/ceres/framework/core/compositions/NetworkConnectivityKt {
-	public static final fun getLocalNetworkConnectivity ()Landroidx/compose/runtime/ProvidableCompositionLocal;
 }
 
 public final class dev/teogor/ceres/framework/core/menu/ComposableSingletons$MenuKt {
